style: 주소가 길어질 경우, 텍스트 말 줄임표 적용 (#42)
